+++ markitup/templatetags/markitup_tags.py
@@ -1,10 +1,14 @@
 from __future__ import unicode_literals
 
 from django import template
-from django.core.urlresolvers import reverse, NoReverseMatch
 from markitup import settings
 from markitup.util import absolute_url
 from markitup.fields import render_func
+
+try:
+    from django.core.urlresolvers import NoReverseMatch, reverse
+except ImportError:
+    from django.urls import NoReverseMatch, reverse
 
 register = template.Library()